+++ linux/filegate_plugin_private.h
@@ -1,3 +1,6 @@
+#ifndef FLUTTER_PLUGIN_FILEGATE_PLUGIN_PRIVATE_H_
+#define FLUTTER_PLUGIN_FILEGATE_PLUGIN_PRIVATE_H_
+
 #include <flutter_linux/flutter_linux.h>
 
 #include "include/filegate/filegate_plugin.h"
@@ -6,5 +9,7 @@
 // https://github.com/flutter/flutter/issues/88724 for current limitations
 // in the unit-testable API.
 
-// Handles the getPlatformVersion method call.
-FlMethodResponse *get_platform_version();
+FlMethodResponse *filegate_get_file_size(FlValue *arguments);
+FlMethodResponse *filegate_pick_files(FlValue *arguments);
+
+#endif  // FLUTTER_PLUGIN_FILEGATE_PLUGIN_PRIVATE_H_